requirements.txt
asgiref
cffi
cryptography
Django
django-cors-headers
django-extensions
djangorestframework
djangorestframework-simplejwt
MarkupSafe
psycopg
pycairo
pycparser
pyglet
PyJWT
pyOpenSSL
sqlparse
tzdata
Werkzeug
